requirements.txt
WTForms==2.3.1
Werkzeug==1.0.1
SQLAlchemy==1.3.1
Pillow==7.1.2
Flask-PageDown==0.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
Flask-Login0.5.0
Flask-Mail==0.9.1
email-validator1.1.1
Flask==1.1.2
bleach==3.1.5